+++ mex/drop/file_history/main.py
@@ -51,9 +51,11 @@ def index() -> rx.Component:
                 ),
             ),
             custom_attrs={"data-testid": "index-card"},
-            style={
-                "width": "100%",
-                "minHeight": "calc(480px * var(--scaling))",
-            },
+            style=rx.Style(
+                {
+                    "width": "100%",
+                    "minHeight": "calc(480px * var(--scaling))",
+                }
+            ),
         ),
     )